add a character modal in combat page if there is no character. Refactored the add character modal in a separate component to reuse
